requirements.txt
<<<<<<< HEAD
fastapi==0.110.0
openai==1.3.7
pinecone-client==2.2.1
pdfminer.six==20221105
python-docx==0.8.11
pydantic==2.4.2
uvicorn==0.27.0
python-dotenv==1.0.1
=======

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==1.10.13
requests==2.32.3
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
PyMuPDF==1.24.7
python-docx==1.1.2
python-multipart==0.0.9
openai==1.35.7
pinecone-client==4.1.1
>>>>>>> 88a2809 (feat: initial FastAPI RAG system)
